penyuaian chat
