+++ src/component/DataChat.js
@@ -1,0 +1,44 @@
+const Messages = [
+    {
+      id: "1",
+      userName: "Clara Ayu Sheila",
+      userImg: require("../../assets/Chat/MUA1.jpg"),
+      messageTime: "4 mins ago",
+      messageText:
+        "Hey there!",
+    },
+    {
+      id: "2",
+      userName: "Natahalief",
+      userImg: require("../../assets/Chat/MUA2.jpg"),
+      messageTime: "2 hours ago",
+      messageText:
+        "Welcome to Artistique",
+    },
+    {
+      id: "3",
+      userName: "Hilga Doui",
+      userImg: require("../../assets/Chat/MUA3.jpg"),
+      messageTime: "1 hours ago",
+      messageText:
+        "Hey there!",
+    },
+    {
+      id: "4",
+      userName: "Fauziah Hanum",
+      userImg: require("../../assets/Chat/MUA4.jpg"),
+      messageTime: "1 day ago",
+      messageText:
+        "Welcome to Artistique",
+    },
+    {
+      id: "5",
+      userName: "Ochi Pramita",
+      userImg: require("../../assets/Chat/MUA5.jpg"),
+      messageTime: "2 days ago",
+      messageText:
+        "Hey there!",
+    },
+  ];
+
+export default Messages;
