+++ src/component/Header.js
@@ -1,38 +1,43 @@
 import React from "react";
-import { View, TouchableOpacity, StatusBar, TextInput } from "react-native";
-import { Box, Text, Input, Flex } from "native-base";
+import { TouchableOpacity} from "react-native";
+import { Box, Input, Flex } from "native-base";
 import Icon from "react-native-vector-icons/Ionicons";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 const Header = (props) => {
   return (
-    <Box>
-      <Flex direction="row">
-        <Input
-          value={props.search}
-          onChangeText={(text) => props.setSearch(text)}
-          placeholder="Search..."
-          backgroundColor={"#FFFFFF"}
-          mt={5}
-          pl={3}
-          rounded={1}
-          flex={1}
-        />
-        <TouchableOpacity
-          style={{
-            justifyContent: "center",
-            alignItems: "center",
-            backgroundColor: "#A01437",
-            marginTop: 20,
-            paddingHorizontal: 20,
-            borderRadius: 5,
-            marginLeft: 10,
-            elevation: 3,
-          }}
-        >
-          <Icon name="search" size={25} color="#FFFFFF" />
-        </TouchableOpacity>
-      </Flex>
-    </Box>
+    <SafeAreaView>
+      <Box >
+        <Flex direction="row">
+          {/* input search */}
+          <Input
+            value={props.search}
+            onChangeText={(text) => props.setSearch(text)}
+            placeholder="Search..."
+            backgroundColor={"#FFFFFF"}
+            // mt={5}
+            pl={3}
+            rounded={1}
+            flex={1}
+          />
+          {/* tombol search */}
+          <TouchableOpacity
+            style={{
+              justifyContent: "center",
+              alignItems: "center",
+              backgroundColor: "#A01437",
+              // marginTop: 20,
+              paddingHorizontal: 20,
+              borderRadius: 5,
+              marginLeft: 10,
+              elevation: 3,
+            }}
+          >
+            <Icon name="search" size={25} color="#FFFFFF" />
+          </TouchableOpacity>
+        </Flex>
+      </Box>
+    </SafeAreaView>
   );
 };
